==> scripts/01.mif_extraction_short.ipynb <==
# Most Informative Frame (MIF) extraction

**Task:**

Having a given set of videos, perform frame-wise classification of these videos using a pre-trained model.
In this case, the model is a ResNet50 pretrained on Moments in Time v1 (MiTv1).

**Total Time duration:** (for ~1.5k MP4s)
* i5-7400 CPU @ 3.00GHz; 16GB RAM; No CUDA: **>15h**
* Google Colab + GPU (CUDA) acceleration: **3-5h**

## Preparation

# Install decord if necessary
# AB: Why? What does it do?
# AB: Generally check out the comments in '01.mif_extraction_exploratory' and copy some infor here as well; this should be sufficient
!pip install --upgrade decord

#### Load pretrained model

#%% Define and load model
from pathlib import Path
import torch
import torch.nn as nn
import torchvision.models as models

with_cuda = False
# Define and load model
from pathlib import Path
import torch
import torch.nn as nn
import torchvision.models as models

with_cuda = False
path_prefix = Path('..')

if with_cuda:
    resnet50 = models.resnet50(pretrained=False, progress=True, num_classes=339).to('cuda')
else:
    resnet50 = models.resnet50(pretrained=False, progress=True, num_classes=339)

# AB: where from?
# Load pretrained weights (MiTv1)
#path_model = Path('/content/drive/MyDrive/resnet50_moments-fd0c4436.pth')
path_model = path_prefix / 'models/weights/resnet50_moments-fd0c4436.pth'
resnet50.load_state_dict(torch.load(path_model))

# AB: source / structure
# Evaluation mode
resnet50.eval()

#### Define transformations

#%% Transformations
import torchvision.transforms as transforms
transformation = transforms.Compose([
                                     transforms.ToPILImage(mode='RGB'), # required if the input image is a nd.array
                                     transforms.Resize(224), # To be changed to rescale to keep the aspect ration?
                                     transforms.CenterCrop((224, 224)), # AB: cropped to 224x224px
                                     transforms.ToTensor(), # AB: creates Tensor (?)
                                     transforms.Normalize(mean=[0.485, 0.456, 0.406],
                                                          std=[0.229, 0.224, 0.225])
])

#### Load categories

# %% Load categories
# AB: How/Where stored? Just the names/labels of the categories? --> make clear
path_labels = path_prefix / 'models/labels/category_momentsv1.txt'

def load_categories():
    """Load categories."""
    with open(path_labels) as f:
        return [line.rstrip() for line in f.readlines()]

# load categories
categories = load_categories()

#### Load videos

#%% Sweep through files in subfolders of path_input
import os
# path_input = path_prefix / 'input_data/MIT_sampleVideos_RAW_final'
path_input = path_prefix / 'input_data/08_MIT_MP4s_25FPS_480x360p_N=834+'

l_videos = []
for path, subdirs, files in os.walk(path_input):
  for name in files:
    if name[-3:] == 'mp4':
      l_videos.append([path.split('/')[-1],   # category
                       name])                 # file name
    else:
      print('Ignored: ', name)

if l_videos:
  l_videos = sorted(l_videos)
print('Total nr. of MP4s: ', len(l_videos))


print(l_videos[:10])

## TRUE class & MIF only

Extract prediction accuracy for the TRUE category on MIFs (most informative frame, i.e. frame w/ highest prediction accuracy)

**Output**

```
pandas.DataFrame w/ columns ['category', 'fname', 'mif_idx', 'softmax[category]']
```

Extract prediction accuracy for the TRUE category on MIFs (most informative frame, i.e. frame w/ highest prediction accuracy)

# %% Sweep through videos
import time
import decord
decord.bridge.set_bridge('native') # Seems to be the fastest option
from decord import cpu, gpu
from decord import VideoReader
from torch.nn import functional as F
import numpy as np
import pandas as pd

start = time.time()

l_mifs = []

# Iterate over entries in l_videos:
for j in range(len(l_videos)):
  # Verbose
  if j%50 == 0:
    print(f'{j}/{len(l_videos)}')
  
  
  # Define path
  category, file_name = l_videos[j]
  cat_idx = categories.index(category)
  path_input_file = str(path_input / category/ file_name)
  
  print(category, cat_idx)
  """
  # Load video with Decord.VideoReader
  vr = VideoReader(path_input_file)
  video_frames = vr.get_batch(range(0, len(vr), 1)).asnumpy()
  
  # Define empty array for accuracies
  pred_accuracies = np.zeros((video_frames.shape[0], ))

  # Iterate over frames
  for i in range(video_frames.shape[0]):
    if with_cuda:
      input = transformation(video_frames[i]).to('cuda')
    else:
      input = transformation(video_frames[i])
    
    # Classification:
    logit = resnet50.forward(input.unsqueeze(0))      # extract output to given input 
    h_x = F.softmax(logit, 1).data.squeeze()[cat_idx] # transform to softmax
    pred_accuracies[i]= h_x

  
  # Append to output list
  l_mifs.append([category, file_name,
                 np.argmax(pred_accuracies),
                 pred_accuracies[np.argmax(pred_accuracies)]])
   """
stop = time.time()
duration = stop-start
print(f'\nTime elapsed:: {duration:.4f}s (~ {duration/j:.2f}s per file)')










# AB: PLACED HERE TO SAVE EARLIER IN BLOCK
import pandas as pd

# Additional information comparing pandas data frame and pickle files see also at the end.

df = pd.DataFrame(l_mifs, columns=['category', 'fname', 'mif_idx', 'softmax[category]'])
print(df)
df.to_csv(path_prefix / 'temp/mifs.csv')

categories[327]

Stack together and save to csv as: \\
  `category, fname, mif_idx`

# import pandas as pd
# 
# # Additional information comparing pandas data frame and pickle files see also at the end.
# 
# df = pd.DataFrame(l_mifs, columns=['category', 'fname', 'mif_idx', 'softmax[category]'])
# print(df)
# df.to_csv(path_prefix / 'temp/mifs.csv')

#### Testing

Using a previously computed dictionary of softmax values, check if same values were collected

from pathlib import Path
import pickle
import numpy as np

# AB: A bit more info considering the pickle file and where it comes from (was this in '01.mif_extraction_exploratory' initially?)

dict_path = path_prefix / 'temp/ResNet50_MiTv1_accuracies_per_category.pkl'
# Load from file
f = open(dict_path, 'rb')
accuracies_per_category = pickle.load(f)

l_categories = categories

# AB: Example ./aiming/yt-0gwUV4Ze-Hs_390.mp4

category_name = 'aiming'
video_fname = 'yt-0gwUV4Ze-Hs_390.mp4'

per_frame_accuracies = np.array(accuracies_per_category[category_name][video_fname])

print(f'\t{video_fname} : Max/Min accuracy at frame:' \
f' {np.argmax(per_frame_accuracies)}/{np.argmin(per_frame_accuracies)}' \
f' with value: {per_frame_accuracies[np.argmax(per_frame_accuracies)]}' \
f' / {per_frame_accuracies[np.argmin(per_frame_accuracies)]}')

df.loc[lambda df: df['fname'] == video_fname]

## ALL classes & ALL frames

**Output**

    ```
    {
        category_i : {
            video_j : [list of per-frame accuracies for all categories]
        }
    }
    ```

# AB: 
# the variable 'd_files_per_category' was not defined in the block below.
# So I added this chunck from an old version of '01.mif_extraction_exploratory.ipynb'
# to make it run.
# Additionally there was a renaming needed of the ResNet variable (see below)

import os

# path_videoset = path_prefix / 'input_data/MIT_sampleVideos_RAW_final_25FPS/'
path_videoset = path_prefix / 'input_data/08_MIT_MP4s_25FPS_480x360p_N=834+' # AB: change path

d_files_per_category = {}
# Iterate over categories in path_prefix
for root, dirs, files in os.walk(path_videoset, topdown=False):
    for name in sorted(dirs):
        category_path = os.path.join(root, name)
        category_name = category_path.split('/')[-1]
        # Iterate over files in cateogory_path
        d_files_per_category[category_name] = {}
        for root2, dirs2, videos  in os.walk(category_path, topdown=False):
            for video in sorted(videos):
                video_fname = os.path.join(root2, video)
                
                d_files_per_category[category_name][video] = video_fname


print(d_files_per_category)

# AB:
# Renaming of variable
resnet50_moments = resnet50

import time
from torch.nn import functional as F

accuracies_per_category = {}

z = 1 # For verbose

time_init = time.time()

# Iterate over categories in path_prefix
for category_name in list(d_files_per_category.keys()):
    print(f'{category_name} {z}/{len(list(d_files_per_category.keys()))}'); z += 1
    
    # Define empty dictionary entry for current category
    accuracies_per_category[category_name] = {}

    # Iterate over files in cateogory_path
    for video in list(d_files_per_category[category_name]):
        # verbose
        print('\t', video)
 
        video_fname = d_files_per_category[category_name][video]
        
        # Load video with Decord.VideoReader
        vr = VideoReader(video_fname, ctx=cpu(0))
        frame_id_list = range(0, len(vr), 1)
        video_frames = vr.get_batch(frame_id_list)

        pred_accuracies = []
        true_category = category_name

        start_time = time.time()
        # Iterate through frames
        for i in range(video_frames.shape[0]):
            # Load frame
            frame = video_frames.asnumpy()[i]

            # Transform
            if with_cuda:
                input = transformation(frame).to('cuda')
            else:
                input = transformation(frame)
            
            # Classification:
            logit = resnet50_moments.forward(input.unsqueeze(0))
            h_x = F.softmax(logit, 1).data.squeeze()
            probs, idx = h_x.sort(0, True)

            # Save accuracies for the all classes (TYPE #2)
            pred_accuracies.append(probs.cpu().numpy())
        
        # Calculate avg duration per frame
        end_time = time.time()
        print('\tAvg duration per frame: %4.4f seconds.' % ((end_time - start_time)/(video_frames.shape[0])))
        
        # Add computed list of accuracies as entry in the output dictionary
        accuracies_per_category[category_name][video] = pred_accuracies

time_final = time.time()
duration = time_final-time_init
print(f'\nTime elapsed: {duration:.2f}s (~ {duration/z:.2f}s per file)')
        

# Save dictionary to pickle file

# Difference: pandas data frame || pickle file
# The pickle (.pkl) file is different form the pandas dataframe used earlier.

# Where the dataframe contains the foruc columns:
# category fname mif_idx  softmax[category]

# the pickle file contains the accuracy per category for all files (massive data)
# additionally it contains the nested structure of the data
# (the mif indices could also restored from the larger pickle file again,
# but not the other way round)

import pickle
dict_path = path_prefix / 'temp/accuracies_per_category_mitv1_fps-25.pkl'

if dict_path.is_file():
    print(f'File {dict_path} already exists. Change dict_path and re-run cell!')
else:
    with open(dict_path, 'wb') as f:
        pickle.dump(accuracies_per_category, f)
        f.close()
    print(f'Saved at {dict_path.absolute()}')

Difference: pandas data frame || pickle file

The pickle (.pkl) file is different form the pandas dataframe used earlier.

Where the dataframe contains the foruc columns:
category fname mif_idx  softmax[category]

the pickle file contains the accuracy per category for all files (massive data) additionally it contains the nested structure of the data
(the mif indices could also restored from the larger pickle file again, but not the other way round)
